move from sage.arith.misc to sage.combinat.q_analogues
